chore: move props to src
