+++ .gitignore
@@ -4,6 +4,7 @@ vars.md
 dev/
 .cache/
 vars.md
+.dev
 
 ### Linux ###
 *~
