--- doc/CONTRIBUTING.md
@@ -1,19 +1,0 @@
-# Contributing
-
-To contribute this project, you need to clone and checkout the `<develop>` branch and then install all dependencies. We prefer [NPM](https://yarnpkg.com/lang/en/) as package manager.
-
-## Development
-
-Run the local development server with autorefresh by running:
-
-```bash
-yarn dev
-```
-
-## Build
-
-You can build all the packages by running this command.
-
-```bash
-yarn build
-```
