+++ package.json
@@ -18,12 +18,12 @@
   },
   "scripts": {
     "clean": "yarn rimraf dist/** .cache",
-    "watch": "parcel demo/index.html --no-source-maps",
+    "watch": "parcel demo/index.html --out-dir ./.dev --no-source-maps",
     "dev": "run-s clean watch",
     "prebuild": "yarn clean",
     "build:css": "parcel build ./src/elements/core/core.pcss --out-dir ./dist -o native-elements.css --no-source-maps",
     "copy:source": "copyfiles -u 2 ./src/elements/**/* ./dist/elements",
-    "copy:props": "copyfiles ./props/**/* ./dist",
+    "copy:props": "copyfiles -u 1 ./src/props/**/* ./dist",
     "build": "run-s build:css",
     "postbuild": "run-s theme copy:source copy:props size",
     "prepublishOnly": "run-s build",
