+++ src/props/core.js
