+++ src/props/details.js
