+++ src/props/dialog.js
