+++ src/props/meter.js
