+++ theme.css
@@ -1,32 +1,6 @@
 :root {
   --accent-color: hsl(220, 100%, 50%);
 
-  --ne-select-border: 1px solid hsl(0, 0%, 89%);
-  --ne-select-hover-border: var(--ne-select-border, 1px solid hsl(0, 0%, 89%));
-  --ne-select-focus-border: var(--ne-select-border, 1px solid var(--accent-color));
-  --ne-select-disabled-border: 1px solid hsl(233, 34%, 96%);
-  --ne-select-background: var(--ne-global-background, hsl(0, 0%, 100%));
-  --ne-select-hover-background: var(--ne-select-background, hsl(0, 0%, 100%));
-  --ne-select-focus-background: var(--ne-select-background, hsl(0, 0%, 100%));
-  --ne-select-disabled-background: hsl(233, 34%, 96%);
-  --ne-select-option-checked-background: var(--accent-color);
-  --ne-select-foreground: var(--ne-global-foreground, hsl(0, 0%, 13%));
-  --ne-select-hover-foreground: var(--ne-select-foreground, hsl(0, 0%, 13%));
-  --ne-select-focus-foreground: var(--ne-select-foreground, hsl(0, 0%, 13%));
-  --ne-select-disabled-foreground: hsl(233, 14%, 72%);
-  --ne-select-option-checked-foreground: hsl(0, 0%, 100%);
-  --ne-select-border-radius: 5px;
-  --ne-select-width: 100%;
-  --ne-select-padding: 0.5em;
-  --ne-select-icon: url(data:image/svg+xml;utf-8,<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24"><path d="M7 8l5-5 5 5m0 8l-5 5-5-5" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" fill="none" stroke="currentColor" stroke-miterlimit="10"/></svg>);
-  --ne-select-icon-size: 1rem;
-  --ne-select-shadow: 0 1px 2px rgba(0, 0, 0, 0.06);
-  --ne-table-head-separator: 2px solid hsl(0, 0%, 89%);
-  --ne-table-row-separator: 1px solid hsl(0, 0%, 89%);
-  --ne-table-striped-background: transparent;
-  --ne-table-background: transparent;
-  --ne-table-cells-y-padding: 16px;
-  --ne-table-cells-x-padding: 16px;
   --ne-textfield-border: 1px solid hsl(0, 0%, 89%);
   --ne-textfield-hover-border: var(--ne-textfield-border, 1px solid hsl(0, 0%, 89%));
   --ne-textfield-focus-border: var(--ne-textfield-border, 1px solid hsl(0, 0%, 89%));
@@ -58,6 +32,4 @@
   --ne-textfield-selection-foreground: hsl(0, 0%, 13%);
   --ne-textfield-resize: vertical;
   --ne-textfield-font-weight: 700;
-  --ne-upload-label-weight: 700;
-  --ne-upload-label-font-size: 1.2rem;
 }
